requirements.txt
snowflake
python-dotenv
streamlit-plotly-events
pandas
python-dotenv